Refactored client in preparation for SSR
