+++ newrelic_agent.log
@@ -35,3 +35,15 @@
 {"v":0,"level":30,"name":"newrelic","hostname":"Nicholass-MBP","pid":80933,"time":"2020-09-16T03:28:22.819Z","msg":"Valid event_harvest_config received. Updating harvest cycles. {\"report_period_ms\":5000,\"harvest_limits\":{\"error_event_data\":8,\"analytic_event_data\":833,\"custom_event_data\":83}}"}
 {"v":0,"level":30,"name":"newrelic","hostname":"Nicholass-MBP","pid":80933,"time":"2020-09-16T03:28:22.821Z","msg":"Agent state changed from connected to started."}
 {"v":0,"level":30,"name":"newrelic","hostname":"Nicholass-MBP","pid":80933,"time":"2020-09-16T03:28:23.826Z","msg":"Starting initial 1000ms harvest."}
+{"v":0,"level":30,"name":"newrelic","hostname":"Nicholass-MBP","pid":82919,"time":"2020-09-17T03:31:58.015Z","msg":"Using New Relic for Node.js. Agent version: 6.13.0; Node version: v12.16.1."}
+{"v":0,"level":40,"name":"newrelic","hostname":"Nicholass-MBP","pid":82919,"time":"2020-09-17T03:31:58.145Z","msg":"The newrelic module must be the first module required.\nThe following modules were required before newrelic and are not being instrumented:\n\texpress: /Users/nzabalza/Immersive/Sprints/SDC/image-service/node_modules/express/index.js\n\tq: /Users/nzabalza/Immersive/Sprints/SDC/image-service/node_modules/qs/lib/index.js\n\tmysql: /Users/nzabalza/Immersive/Sprints/SDC/image-service/node_modules/mysql/index.js\n\tbluebird: /Users/nzabalza/Immersive/Sprints/SDC/image-service/node_modules/bluebird/js/release/bluebird.js"}
+{"v":0,"level":30,"name":"newrelic","hostname":"Nicholass-MBP","pid":82919,"time":"2020-09-17T03:31:58.146Z","msg":"Agent state changed from stopped to starting."}
+{"v":0,"level":30,"name":"newrelic","hostname":"Nicholass-MBP","pid":82919,"time":"2020-09-17T03:31:58.147Z","msg":"Not adding native metric sampler.","component":"sampler","error":{"message":"Cannot find module '@newrelic/native-metrics'\nRequire stack:\n- /Users/nzabalza/Immersive/Sprints/SDC/image-service/node_modules/newrelic/lib/sampler.js\n- /Users/nzabalza/Immersive/Sprints/SDC/image-service/node_modules/newrelic/lib/agent.js\n- /Users/nzabalza/Immersive/Sprints/SDC/image-service/node_modules/newrelic/index.js\n- /Users/nzabalza/Immersive/Sprints/SDC/image-service/server/index.js\n- /Users/nzabalza/Immersive/Sprints/SDC/image-service/server/start.js","stack":"Error: Cannot find module '@newrelic/native-metrics'\nRequire stack:\n- /Users/nzabalza/Immersive/Sprints/SDC/image-service/node_modules/newrelic/lib/sampler.js\n- /Users/nzabalza/Immersive/Sprints/SDC/image-service/node_modules/newrelic/lib/agent.js\n- /Users/nzabalza/Immersive/Sprints/SDC/image-service/node_modules/newrelic/index.js\n- /Users/nzabalza/Immersive/Sprints/SDC/image-service/server/index.js\n- /Users/nzabalza/Immersive/Sprints/SDC/image-service/server/start.js\n    at Function.Module._resolveFilename (internal/modules/cjs/loader.js:982:15)\n    at Function.wrappedResolveFilename [as _resolveFilename] (/Users/nzabalza/Immersive/Sprints/SDC/image-service/node_modules/newrelic/lib/shimmer.js:364:42)\n    at Function.Module._load (internal/modules/cjs/loader.js:864:27)\n    at Function.wrappedLoad [as _load] (/Users/nzabalza/Immersive/Sprints/SDC/image-service/node_modules/newrelic/lib/shimmer.js:373:24)\n    at Module.require (internal/modules/cjs/loader.js:1044:19)\n    at require (internal/modules/cjs/helpers.js:77:18)\n    at Object.start (/Users/nzabalza/Immersive/Sprints/SDC/image-service/node_modules/newrelic/lib/sampler.js:168:30)\n    at Agent.start (/Users/nzabalza/Immersive/Sprints/SDC/image-service/node_modules/newrelic/lib/agent.js:212:11)\n    at createAgent (/Users/nzabalza/Immersive/Sprints/SDC/image-service/node_modules/newrelic/index.js:152:9)\n    at initialize (/Users/nzabalza/Immersive/Sprints/SDC/image-service/node_modules/newrelic/index.js:80:15)\n    at Object.<anonymous> (/Users/nzabalza/Immersive/Sprints/SDC/image-service/node_modules/newrelic/index.js:37:3)\n    at Module._compile (internal/modules/cjs/loader.js:1158:30)\n    at Object.Module._extensions..js (internal/modules/cjs/loader.js:1178:10)\n    at Module.load (internal/modules/cjs/loader.js:1002:32)\n    at Function.Module._load (internal/modules/cjs/loader.js:901:14)\n    at Module.require (internal/modules/cjs/loader.js:1044:19)"}}
+{"v":0,"level":30,"name":"newrelic","hostname":"Nicholass-MBP","pid":82919,"time":"2020-09-17T03:31:58.147Z","msg":"Starting New Relic for Node.js connection process."}
+{"v":0,"level":30,"name":"newrelic","hostname":"Nicholass-MBP","pid":82919,"time":"2020-09-17T03:31:58.148Z","msg":"Agent state changed from starting to connecting."}
+{"v":0,"level":30,"name":"newrelic","hostname":"Nicholass-MBP","pid":82919,"time":"2020-09-17T03:31:59.926Z","msg":"Agent state changed from connecting to connected."}
+{"v":0,"level":30,"name":"newrelic","hostname":"Nicholass-MBP","pid":82919,"time":"2020-09-17T03:31:59.926Z","msg":"Connected to collector-003.newrelic.com:443 with agent run ID BbtNifKQB8b2ACvVkigfMZxfYtlkAAgBAAAnIQEAAUPnAgQoHzGaAwAGNi4xMy4wAA1OaWNob2xhc3MtTUJQABFTREMtSW1hZ2UtU2VydmljZQ.","component":"collector_api"}
+{"v":0,"level":30,"name":"newrelic","hostname":"Nicholass-MBP","pid":82919,"time":"2020-09-17T03:31:59.926Z","msg":"Reporting to: https://rpm.newrelic.com/accounts/2872722/applications/673132954","component":"collector_api"}
+{"v":0,"level":30,"name":"newrelic","hostname":"Nicholass-MBP","pid":82919,"time":"2020-09-17T03:31:59.928Z","msg":"Valid event_harvest_config received. Updating harvest cycles. {\"report_period_ms\":5000,\"harvest_limits\":{\"error_event_data\":8,\"analytic_event_data\":833,\"custom_event_data\":83}}"}
+{"v":0,"level":30,"name":"newrelic","hostname":"Nicholass-MBP","pid":82919,"time":"2020-09-17T03:31:59.930Z","msg":"Agent state changed from connected to started."}
+{"v":0,"level":30,"name":"newrelic","hostname":"Nicholass-MBP","pid":82919,"time":"2020-09-17T03:32:00.931Z","msg":"Starting initial 1000ms harvest."}
