+++ webpack.config.js
@@ -5,17 +5,19 @@ const CompressionPlugin = require('compression-webpack-plugin');
 
 
 module.exports = {
-  entry: `${SRC_DIR}/index.jsx`,
+  entry: {
+    'bundle.js': `${SRC_DIR}/entryCSR.jsx`
+  },
   output: {
-    filename: 'bundle.js',
+    filename: '[name]',
     path: DIST_DIR
   },
   plugins: [
     new CompressionPlugin({
-        filename: '[path].gz[query]',
-        algorithm: 'gzip',
-        test: /\.js$/,
-        threshold: 5000,
+      filename: '[path].gz[query]',
+      algorithm: 'gzip',
+      test: /\.js$/,
+      threshold: 5000,
     }),
   ],
   module : {
@@ -31,8 +33,8 @@ module.exports = {
     ]
   },
   externals: {
-   'react': 'React',
-   'react-dom' : 'ReactDOM',
-   'jquery': 'jQuery'
+    'react': 'React',
+    'react-dom' : 'ReactDOM',
+    'jquery': 'jQuery'
   }
 };
